# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: NinerLog API
  version: 1.0.0
  description: |
    RESTful API for EASA/FAA compliant pilot logbook management.
    
    ## Features
    - Multi-license support (PPL, SPL, CPL, ATPL, IR)
    - EASA and FAA regulatory compliance
    - Flight logging with detailed time tracking
    - Currency calculations
    - Statistics and hour totals per license
    
    ## Authentication
    All endpoints except `/auth/register` and `/auth/login` require JWT authentication.
    Include the token in the Authorization header: `Bearer <token>`
    
  contact:
    name: Frederic Jaeckel
    email: frederic@ninerlog.app
  license:
    name: TBD
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://app.ninerlog.com/api/v1
    description: Production server
...
paths:
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: List all users (privacy-preserving)
      description: |
        Returns a paginated list of all user accounts with metadata only.
        No flight data, credentials, or personal documents are exposed.
        Requires admin privileges.
      operationId: listAdminUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
          description: Search by email or name
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAdminUsers'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{userId}:
    delete:
      tags:
        - Admin
      summary: Permanently delete a user account
      description: |
        Permanently delete a user account and ALL their content (flights, aircraft,
        licenses, contacts, credentials, backups, notifications, etc.) via cascading
        deletes. This action is irreversible. The deletion is recorded in the admin
        audit log (with the target user reference set to NULL after the cascade).
        An admin cannot delete their own account.
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User and all their content deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{userId}/disable:
    post:
      tags:
        - Admin
      summary: Disable a user account
      description: Disable a user account and revoke all tokens. Logged to audit trail.
      operationId: disableUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User disabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{userId}/enable:
    post:
      tags:
        - Admin
      summary: Re-enable a user account
      description: Re-enable a previously disabled user account. Logged to audit trail.
      operationId: enableUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User enabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{userId}/unlock:
    post:
      tags:
        - Admin
      summary: Unlock a locked user account
      description: Clear brute-force lockout for a user account. Logged to audit trail.
      operationId: unlockUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User unlocked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{userId}/reset-2fa:
    post:
      tags:
        - Admin
      summary: Reset 2FA for a user
      description: Disable 2FA for a user who lost their authenticator. Logged to audit trail.
      operationId: resetUser2fa
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 2FA reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/refresh
...
  schemas:
...
    AdminUser:
      type: object
      required:
        - id
        - email
        - name
        - createdAt
        - emailVerified
        - twoFactorEnabled
        - disabled
        - flightCount
        - aircraftCount
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          description: Last successful login timestamp
        emailVerified:
          type: boolean
          description: Whether the user has verified their email address and completed sign-up
        twoFactorEnabled:
          type: boolean
        disabled:
          type: boolean
          description: Whether the account is disabled by admin
        locked:
          type: boolean
          description: Whether the account is locked due to brute-force protection
        lockedUntil:
          type: string
          format: date-time
          description: When the brute-force lock expires
        flightCount:
          type: integer
          description: Total number of flights logged
        aircraftCount:
          type: integer
          description: Number of aircraft in database
    PaginatedAdminUsers:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'
        pagination:
          type: object
          required:
            - page
            - pageSize
            - total
            - totalPages
          properties:
            page:
              type: integer
            pageSize:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer

    # ---------------- Cloud Backups -----------------------------------------
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Validation failed
          description: Error message
        code:
          type: string
          example: email_not_verified
          description: |
            Optional machine-readable error code, used by clients to distinguish
            between conditions that share an HTTP status (e.g. `email_not_verified`
            vs `account_disabled` for 403 on /auth/login).
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: totalTime
              message:
                type: string
                example: Must be greater than 0
          description: Detailed validation errors
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation failed
            details:
              - field: totalTime
                message: Must be greater than 0
              - field: nightTime
                message: Cannot exceed total time
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
    Forbidden:
      description: Forbidden - admin access required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Admin access required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Resource not found
...